pyproject.toml
[project]
name = "django-ninja-search"
version = "0.1.0"
description = "Adds a descorator for searching and filtering in django-ninja views"
authors = [
    {name = "Anand R Nair",email = "anand547@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.3,<6.0.0)",
    "django-ninja (>=1.4.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"

